+++ lib/macros4cuke/macro-collection.rb
@@ -31,14 +31,14 @@ public
   # Pre-condition: there is no existing macro with the same key.
   # [aPhrase] The text that is enclosed between the square brackets.
   # [aTemplate] A text that consists of a sequence of Cucumber steps.
-  # These steps
-  def add_macro(aPhrase, aTemplate)
-    new_macro = MacroStep.new(aPhrase, aTemplate)
+  # [useTable] A boolean that indicates whether a table should be used to pass actual values.
+  def add_macro(aPhrase, aTemplate, useTable)
+    new_macro = MacroStep.new(aPhrase, aTemplate, useTable)
     
     # Prevent collision of macros (macros with same phrase).
     # This can occur if a macro was defined in a background section.
     # An exception is raised if the phrase syntax of both macros are the     
-    raise DuplicateMacroError.new(aPhrase) if find_macro(aPhrase)
+    raise DuplicateMacroError.new(aPhrase) if find_macro(aPhrase, useTable)
     
     @macro_steps[new_macro.key] = new_macro    
 
@@ -52,8 +52,8 @@ public
   # Each couple is of the form: argument name, a value.
   # Multiple rows with same argument name are acceptable.  
   def render_steps(aPhrase, rawData = nil)
-
-    macro = find_macro(aPhrase) 
+    useTable = ! rawData.nil?
+    macro = find_macro(aPhrase, useTable) 
     raise UnknownMacroError.new(aPhrase) if macro.nil?
 
     # Render the steps
@@ -62,8 +62,8 @@ public
 
 private
   # Retrieve the macro, given a phrase.
-  def find_macro(aMacroPhrase)
-    return @macro_steps[MacroStep::macro_key(aMacroPhrase, :invokation)]
+  def find_macro(aMacroPhrase, useTable)
+    return @macro_steps[MacroStep::macro_key(aMacroPhrase, useTable, :invokation)]
   end
 
 end # class